facdb tweaks following ingest migration (#1471)
* fixup ingest multiple gha

* revert certain #1313 tweaks

* clean up facdb recipe, add comments for csv datasets

* facdb pipeline tweaks for ingest

* convert nysdec_solidwaste to 4326

* stop pinning council districts to 22c

* restructure facdb_boro query

* use ingest for dcp_facilities_with_unmapped

+++ products/facilities/facdb/sql/_qaqc.sql
@@ -125,7 +125,7 @@ WITH
 new AS (
     SELECT
         captype,
-        sum(capacity) AS sum_new
+        sum(capacity::numeric)::integer AS sum_new
     FROM facdb
     GROUP BY captype
 ),
